Renamed index.js to index.tsx
