+++ package.json
@@ -40,6 +40,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^17.0.13",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
     "eslint-config-airbnb": "^19.0.4",
