+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './components/App';
+import App from './App.js';
 import "bootstrap/dist/css/bootstrap.min.css";
 
 ReactDOM.render(
